Add RequestId as `id` for Error
To help make errors more traceable from a user's perspective we
would like to be able for the api user to issue the id of the
error they got when debugging so we could then look through our
logs for this error.

This uses the built in `Plug.RequestId` plug that ships with
Phoenix to help bring that id forward.  It is overrideable via
specifying the `request_id` and using the same token that is
set for the aforementioned plug.